// src/types/PairInfo.ts
export interface PairInfo {
  pair: string;
  base: string;
  quote: string;
  basedisplaydecimals: number;
  quotedisplaydecimals: number;
  baseaddress: string;
  quoteaddress: string;
  mintrade_amnt: string;
  maxtrade_amnt: string;
  base_evmdecimals: number;
  quote_evmdecimals: number;
  auctionmode: number;
  auctionendtime: string;
  status: string;
}